user side image uploading issue fixed

+++ client/src/Components/Header/Header.jsx
@@ -4,53 +4,57 @@ import { useSelector, useDispatch } from 'react-redux';
 import { change } from '../../Redux/usernameReducer';
 import { changeImage } from '../../Redux/userImageReducer';
 import axios from '../../utils/axios'
-import {verifyToken}from '../../utils/Constants'
-
-
-import jwt_decode from 'jwt-decode'
-
+import { verifyToken } from '../../utils/Constants'
+import Swal from "sweetalert2";
 
 function Header() {
-
-
     const navigate = useNavigate();
     const dispatch = useDispatch()
-  
     useEffect(() => {
         const Token = localStorage.getItem("token");
         if (Token) {
             axios.post(verifyToken, JSON.stringify({ Token }), { headers: { "Content-Type": "application/json" } }).then((res) => {
-               
+
                 dispatch(change(res.data.user.username))
-                console.log(res.data.user.image);
+             
                 dispatch(changeImage(res.data.user.image))
-                console.log(res,"resssssssssss");
+               
             }).catch((err) => {
                 localStorage.removeItem('adminToken');
 
             })
 
         }
-    }, []);
-
-
+    }, [dispatch]);
 
     const username1 = useSelector((state) => {
         return state.username;
     })
     const userImage = useSelector((state) => {
         return state.userImage;
-        
+
     })
-    console.log(username1,"666666666666666");
-    
-
-
-
-
+   
     const logout = () => {
-        localStorage.clear();
-        dispatch({type:'logout'})
+        Swal.fire({
+            title: 'Logout?',
+            text: "You won't be able to revert this!",
+            icon: 'warning',
+            showCancelButton: true,
+            confirmButtonColor: '#3085d6',
+            cancelButtonColor: '#d33',
+            confirmButtonText: 'Logout'
+        }).then((result) => {
+            if (result.isConfirmed) {
+                localStorage.clear();
+                dispatch({ type: 'logout' })
+                navigate('/')
+            }
+        })
+
+
+
+
     };
 
 
@@ -76,7 +80,7 @@ function Header() {
                     </button>
                     {/* Collapsible wrapper */}
                     <div className="collapse navbar-collapse" id="navbarSupportedContent">
-                        {/* Navbar brand */}
+                        {/* Navbar brand
                         <Link className="navbar-brand mt-2 mt-lg-0" to={'/'}>
                             <img
                                 src="https://mdbcdn.b-cdn.net/img/logo/mdb-transaprent-noshadows.webp"
@@ -85,14 +89,25 @@ function Header() {
                                 alt="MDB Logo"
                                 loading="lazy"
                             />
-                        </Link>
+                        </Link> */}
                         {/* Left links */}
                         <ul className="navbar-nav me-auto mb-2 mb-lg-0">
                             <li className="nav-item">
                                 <Link className="nav-link" to={'/'}>
-                                    Home
+                                     <h4>Home</h4>
                                 </Link>
                             </li>
+                            <li className="nav-item">
+                                <Link className="nav-link" to={'/page1'}>
+                                     <h4>Page 1</h4>
+                                </Link>
+                            </li>
+                            <li className="nav-item">
+                                <Link className="nav-link" to={'/page2'}>
+                                     <h4>Page 2</h4>
+                                </Link>
+                            </li>
+                            
                             {/* <li className="nav-item">
                                 <a className="nav-link" href="#">
                                     Team
@@ -110,10 +125,10 @@ function Header() {
                     {/* Right elements */}
                     <div className="d-flex align-items-center">
                         {/* Icon */}
-                        {username1 ?   <Link className="text-reset me-3" onClick={logout} to={'/'}> Logout</Link>:<Link className="text-reset me-3" to={'/login'}>LogIn</Link> }
+                        {username1 ? <button className="text-reset me-3 btn  btn-outline-danger btn-sm" onClick={logout} > Logout</button> : <Link className=" text-reset me-3" to={'/login'}>LogIn</Link>}
                         <div className="dropdown">
                             <Link
-                                className="dropdown-toggle d-flex align-items-center hidden-arrow"
+                                className="dropdown-toggle d-flex align-items-center hidden-arrow text-decoration-none text-dark"
                                 to={'/profile'}
                                 id="navbarDropdownMenuAvatar"
                                 role="button"
@@ -127,7 +142,7 @@ function Header() {
                                     className="rounded-circle"
                                     height={25}
                                     alt="Black and White Portrait of a Man"
-                                    loading="lazy"
+                                    // loading="lazy"
                                 />
                             </Link>
                             <ul
